requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.29.0
orjson>=3.9.10
python-dotenv>=1.0.1
requests>=2.31.0